zar: initial support of s3
Use pkg iosrc when opening/creating files within zar, allowing zar
directories to be accessed via origins other than the local filesystem.

Also:
- iosrc.URI: A struct for unifying file paths and url paths so that
operations like AppendPath will work for all supported path types.

+++ cmd/zar/rm/command.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/brimsec/zq/archive"
 	"github.com/brimsec/zq/cmd/zar/root"
+	"github.com/brimsec/zq/pkg/iosrc"
 	"github.com/mccanne/charm"
 )
 
@@ -29,12 +30,14 @@ func init() {
 
 type Command struct {
 	*root.Command
-	root string
+	root          string
+	relativePaths bool
 }
 
 func New(parent charm.Command, f *flag.FlagSet) (charm.Command, error) {
 	c := &Command{Command: parent.(*root.Command)}
 	f.StringVar(&c.root, "R", os.Getenv("ZAR_ROOT"), "root directory of zar archive to walk")
+	f.BoolVar(&c.relativePaths, "relative", false, "display paths relative to root")
 	return c, nil
 }
 
@@ -61,18 +64,33 @@ func (c *Command) Run(args []string) error {
 		return err
 	}
 
-	return archive.Walk(ark, func(zardir string) error {
+	return archive.Walk(ark, func(zardir iosrc.URI) error {
+		if zardir.Scheme != "file" {
+			return errors.New("only file paths currently supported for this command")
+		}
+		dir := zardir.Filepath()
 		for _, name := range args {
-			path := filepath.Join(zardir, name)
+			path := filepath.Join(dir, name)
 			if fileExists(path) {
 				if err := os.Remove(path); err != nil {
 					return err
 				}
-				fmt.Printf("%s: removed\n", path)
+				fmt.Printf("%s: removed\n", c.printable(ark.DataPath.Filepath(), path))
 			} else {
-				fmt.Printf("%s: not found\n", path)
+				fmt.Printf("%s: not found\n", c.printable(ark.DataPath.Filepath(), path))
 			}
 		}
 		return nil
 	})
+}
+
+func (c *Command) printable(root, path string) string {
+	if c.relativePaths {
+		p, err := filepath.Rel(root, path)
+		if err != nil {
+			panic(err)
+		}
+		path = p
+	}
+	return path
 }
